move `query_and_resolve_tree_transform_at_entity`
